Refactor: 프로필 수정 관련 파일 폴더 이동
