+++ FE/src/components/modal/profile_modal/updateProfile/updateNicknameListener.js
@@ -1,38 +1,5 @@
-import apiEndpoints from '../../../constants/apiConfig.js';
-import { getCookie } from '../../../utils/getCookie.js';
-import { nicknameValidCheck } from '../../../pages/register/registerValidCheck.js';
-
-async function updateProfileAPI(value) {
-	const csrfToken = getCookie('csrftoken');
-	const formData = new FormData();
-
-	if (typeof value === 'string') {
-		formData.append('nickname', value);
-	}
-
-	try {
-		const response = await fetch(apiEndpoints.UPDATE_USER_URL, {
-			method: 'PATCH',
-			headers: {
-				'X-CSRFToken': csrfToken
-			},
-			body: formData
-		});
-
-		const data = await response.json();
-		const { status, ok } = response;
-
-		if (ok) {
-			return data.nickname;
-		}
-		if (status === 400) {
-			return null;
-		}
-	} catch (error) {
-		console.error('Error:', error);
-	}
-	return null;
-}
+import { nicknameValidCheck } from '../../../../pages/register/registerValidCheck.js';
+import { updateProfileAPI } from './updateProfileAPI.js';
 
 function toggleNicknameEditUI(isEditing, elements, nickname = '') {
 	const { modifyNicknameButton, modifyCancelButton, input, span } = elements;
